Handle edge race condition
Prior to this change, it was possible to have a race condition (to which
I didn't fully understand). We need to await the .abort() and .retry()
however this also introduces an unhandled promise because Embre throws
TransitionAborted errors which you normally don't see as Ember handles
this in the default willTransition().